Fixed rest scheme, added gson converter to retrofit

+++ app/src/main/java/com/vmarienkov/beers/data/network/Response.kt
@@ -7,11 +7,11 @@ data class Response(
     val brewers_tips: String?,
     val contributed_by: String?,
     val description: String,
-    val ebc: Int?,
+    val ebc: Double?,
     val first_brewed: String?,
     val food_pairing: List<String?>?,
-    val ibu: Int?,
-    val id: Int,
+    val ibu: Double?,
+    val id: Long,
     val image_url: String?,
     val ingredients: Ingredients?,
     val method: Method?,
@@ -19,8 +19,8 @@ data class Response(
     val ph: Double?,
     val srm: Double?,
     val tagline: String?,
-    val target_fg: Int?,
-    val target_og: Int?,
+    val target_fg: Double?,
+    val target_og: Double?,
     val volume: Volume?
 )
 
@@ -36,22 +36,22 @@ data class Fermentation(
 
 data class Temp(
     val unit: String?,
-    val value: Int?
+    val value: Double?
 )
 
 data class MashTemp(
-    val duration: Int?,
+    val duration: Double?,
     val temp: Temp?
 )
 
 data class Volume(
     val unit: String?,
-    val value: Int?
+    val value: Double?
 )
 
 data class BoilVolume(
     val unit: String?,
-    val value: Int?
+    val value: Double?
 )
 
 data class Ingredients(
@@ -79,5 +79,5 @@ data class Hop(
 
 data class AmountX(
     val unit: String?,
-    val value: Int?
+    val value: Double?
 )
