+++ app/src/main/java/com/vmarienkov/beers/domain/model/Beer.kt
@@ -1,3 +1,3 @@
 package com.vmarienkov.beers.domain.model
 
-data class Beer(val id: Int, val photoUrl: String?, val desc: String)
+data class Beer(val id: Long, val photoUrl: String?, val desc: String)
